Merge branch 'phy-unbind-crash'
Florian Fainelli says:

====================
net: phy: Fix PHY unbind crash

This fixes crashes when the PHY driver is no longer bound to the device.

There is still a fair amount of work to be done to get the unbind -> bind
sequent to result in a functional state, but that will be net-next material.

These two problems existed for as long as PHYLIB as been around.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 5ed3fea0d6ffcef494b111ade0910a4ff149c331 ]