Merge branch 'qed-bug-fixes'
Yuval Mintz says:

====================
qed*: Bug fixes

This is a respin of the series with same name for net-next
instead of net [only patch #10 is changed].

The 3 first patches here are a repost of
("qed: RoCE infrastructure fixes"). The fourth is an additional
RoCE-related infrastructure fix, and the latter contain various fixes
to qed/qede.

Please consider applying these to `net-next'.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 5a4dcdcac8e591e235943998caffdf6ac421a422 ]