Merge branch 'bpf-unlocking-fix'
Daniel Borkmann says:

====================
BPF fix with regards to unlocking

This set fixes the issue Eric was reporting recently [1].
First patch is a prerequisite discussed with Laura that is
needed for the later fix in the second one. I've tested this
extensively and it does not reproduce anymore on my side
after the fix. Thanks & sorry about that!

  [1] https://www.spinics.net/lists/netdev/msg421877.html
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: d54fef315399e0b16f8ae2b41167f34f8df12e88 ]