dt-bindings: gemini: augment Gemini bindings to reflect Faraday origin
When we merged the Gemini interrupt controller it was not yet
discovered that this IP block is actually a standard Faraday
Technology interrupt controller.

As the IP block will probably appear in other designs as well,
let's augment the DT bindings to reflect that it is first and
foremost a standard Faraday part with a function name (FTINTC010)
so that people reusing the IP easily find the driver they need.

Sorry for the mistakes due to lack of information.

Cc: Greentime Hu <green.hu@gmail.com>
Acked-by: Rob Herring <robh@kernel.org>
Acked-by: Hans Ulli Kroll <ulli.kroll@googlemail.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>

[ upstream commit: 77a452551b73aec0817d58ecf6d0dd1e74dab90c ]