Merge branch 'r8152-rx-settings'
Hayes Wang says:

====================
r8152: fix the rx settings of RTL8153

The RMS and the rx early size should base on the same rx size. However,
the RMS is set to 9K bytes now and the rx early depends on mtu. For using
the rx buffer effectively, sync the two settings according to the mtu.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 8fb106b28c68d6c81cf6d69773b8cfdbf657b94e ]