Merge branch 'ibmvnic-cleanup-resource-handling'
Nathan Fontenot says:

====================
ibmvnic: Cleanup resource handling

In order to better manage the resources of the ibmvnic driver, this set of
patches creates a set of initialization and release routines for the
drivers resources. Additionally, some patches do some re-naming of the
affected routines so that there is a common naming scheme in the driver.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 7801a3225e7fe05f7601ab42a5669ba633c4df54 ]