Merge branch 'ftgmac-rework-batch2-rx-path'
Benjamin Herrenschmidt says:

====================
ftgmac: Rework batch 2 - RX path

This is the second batch of updates to the ftgmac100 driver.

This one tackles the RX path of the driver, simplifying
it greatly to match common practice while significantly
increasing the performance.

(The bulk of the performance gains of my series will be
provided by the TX path improvements, notably fragmented
sends, these will be in the next batch).
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: dc423b6be10c56d0dd957caf307f12d16f728f07 ]