ARM: dts: dra7x-evm: Enable dual-role mode for USB1
USB1 port is micro-AB type and can function as peripheral
as well as host. Enable dual-role mode for USB1.

We don't want to use the OTG controller block on this
platform as it limits host mode to high-speed. Instead
we rely on extcon framework to give us ID events for
dual-role mode detection.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: a6f627e24139dd8f60057f4812f3fc573c2ad2e3 ]